requirements.txt
Flask==2.0.1
pyttsx3==2.90
Werkzeug==2.0.1
gunicorn==21.2.0
python-docx==0.8.11